+++ Dragon/include/operators/vision/roi_align_op.h
@@ -22,10 +22,10 @@ class ROIAlignOp final : public Operator<Context> {
  public:
     ROIAlignOp(const OperatorDef& def, Workspace *ws)
         : Operator<Context>(def, ws),
-          pool_h(OperatorBase::Arg<int>("pool_h", 0)),
-          pool_w(OperatorBase::Arg<int>("pool_w", 0)),
+          pool_h(OperatorBase::Arg<int64_t>("pool_h", 0)),
+          pool_w(OperatorBase::Arg<int64_t>("pool_w", 0)),
           spatial_scale(OperatorBase::Arg<float>("spatial_scale", 1.f)),
-          sampling_ratio(OperatorBase::Arg<int>("sampling_ratio", 2)) {
+          sampling_ratio(OperatorBase::Arg<int64_t>("sampling_ratio", 2)) {
         CHECK_GT(pool_h, 0) << "\npool_h must > 0";
         CHECK_GT(pool_w, 0) << "\npool_w must > 0";
     }
@@ -44,10 +44,10 @@ class ROIAlignGradientOp final : public Operator<Context> {
  public:
     ROIAlignGradientOp(const OperatorDef& def, Workspace *ws)
         : Operator<Context>(def, ws),
-          pool_h(OperatorBase::Arg<int>("pool_h", 0)),
-          pool_w(OperatorBase::Arg<int>("pool_w", 0)),
+          pool_h(OperatorBase::Arg<int64_t>("pool_h", 0)),
+          pool_w(OperatorBase::Arg<int64_t>("pool_w", 0)),
           spatial_scale(OperatorBase::Arg<float>("spatial_scale", 1.f)),
-          sampling_ratio(OperatorBase::Arg<int>("sampling_ratio", 2)) {
+          sampling_ratio(OperatorBase::Arg<int64_t>("sampling_ratio", 2)) {
         CHECK_GT(pool_h, 0) << "\npool_h must > 0";
         CHECK_GT(pool_w, 0) << "\npool_w must > 0";
     }
